refactor: move currentContentIndex to contentNavigation component
